# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs",
        "version": "1"
    },
    "paths": {
...
        "/api/v1/driver/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Данные о заказах водителя",
                "description": "\n            Необходим токен авторизации.\n            Возвращает все заказы водителя (которые он выполнял).",
                "operationId": "ebee642cc7e23feab39fc015fc55ae07",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "11"
                                                },
                                                "status": {
                                                    "properties": {
                                                        "key": {
                                                            "type": "string",
                                                            "example": "active"
                                                        },
                                                        "value": {
                                                            "type": "string",
                                                            "example": "Активный"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "type": {
                                                    "properties": {
                                                        "key": {
                                                            "type": "string",
                                                            "example": "tow_truck"
                                                        },
                                                        "value": {
                                                            "type": "string",
                                                            "example": "Эвакуатор"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/driver/orders/{order}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Данные о заказе водителя",
                "description": "\n            Необходим токен авторизации (водителя).\n            Возвращает данные заказа водителя (который он выполнял).",
                "operationId": "3102b9c6191f2074224a951a59ccf7c5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "11"
                                                },
                                                "driver_id": {
                                                    "type": "integer|null",
                                                    "example": "87091160"
                                                },
                                                "date_start": {
                                                    "type": "string",
                                                    "example": "2024-09-28 21:30:46"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "tow_truck"
                                                },
                                                "client_comment": {
                                                    "type": "string",
                                                    "example": "Some comments here..."
                                                },
                                                "location_start": {
                                                    "properties": {
                                                        "lat": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": "55.751244"
                                                        },
                                                        "lot": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": "37.618423"
                                                        },
                                                        "address": {
                                                            "type": "string",
                                                            "example": "19402 Langworth Crest\\nSchneidermouth, WV 51672"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/driver/orders/{order}/take": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Взять заказ в работу (для водителей)",
                "description": "\n            Необходим токен авторизации (водителя).\n            Взять заказ в работу (для активированного и не занятого другим заказом водителя).",
                "operationId": "0bea4ad16db89f2e28860636eb251b3c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "11"
                                                },
                                                "client": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 60081201
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Marjorie Okuneva"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "79820354831"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "client"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "dahlia81@example.org"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "В процессе"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Эвакуатор"
                                                },
                                                "client_comment": {
                                                    "type": "string",
                                                    "example": "Some comments here..."
                                                },
                                                "location_start": {
                                                    "properties": {
                                                        "lat": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": "55.751244"
                                                        },
                                                        "lot": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": "37.618423"
                                                        },
                                                        "address": {
                                                            "type": "string",
                                                            "example": "19402 Langworth Crest\\nSchneidermouth, WV 51672"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "date_start": {
                                                    "type": "string",
                                                    "example": "2024-10-10 23:21:40"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}